backend/pyproject.toml
[tool.poetry]
name = "test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
yt-dlp = "^2023.1.2"
pycryptodome = "^3.16.0"
django = "^4.1.5"
firebase-admin = "^6.0.1"
mailjet-rest = "^1.3.4"
requests = "^2.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
